fix: make the page turn actually reach the page
Four independent defects each broke page turning on their own. They were
stacked, so fixing one only exposed the next.

The WebGL2 program never linked. `u_pass` is declared in both stages and relied
on the stage default precision, which is `highp` in a vertex shader and
`mediump` in a fragment shader, so every physical turn failed with
"Precisions of uniform 'u_pass' differ between VERTEX and FRAGMENT shaders" and
disabled the controller.

The controller was torn down by its own teardown effect. The effect listed
`state` in its dependencies, so React ran the cleanup on every phase change and
the transition into `preparing` cancelled the turn that produced it.

The texture cache retired a generation permanently. The surface effect releases
its generation on cleanup and can legitimately re-run for the same turn, and
every retry then answered `stale`. Preparing a generation that is still the
latest now un-retires it.

The surface re-reports readiness on each re-render, and starting a turn resets
its animation clock, so the turn rewound continuously and never landed. A turn
generation is now started once.

Separately, an automatic backward turn derived its grab corner and destination
from the reading direction alone. It therefore swept away from its destination,
the release read as a cancel, and the previous-page control never moved the
reader in either reading direction.

Adds guards for each: a shader test asserting shared uniforms carry matching
explicit precision, a texture-cache test for re-preparing a released
generation, and hook tests for surviving the starting re-render, ignoring
repeated readiness, and committing a backward turn. Also updates the RTL visual
scenario, which still asserted the `committing` phase name from the turn engine
that was replaced.

+++ src/rendering/pageTurn/webgl2Shaders.ts
@@ -9,7 +9,7 @@ uniform int u_control_columns;
 uniform int u_control_rows;
 uniform float u_progress;
 uniform vec2 u_grab_point;
-uniform int u_pass;
+uniform highp int u_pass;
 
 in vec3 a_position;
 in vec2 a_uv;
@@ -85,7 +85,7 @@ uniform sampler2D u_under;
 uniform sampler2D u_stationary;
 uniform float u_luminance_min;
 uniform float u_luminance_max;
-uniform int u_pass;
+uniform highp int u_pass;
 
 in vec2 v_uv;
 in vec3 v_normal;
